+++ packages/design-system/src/components/web-components/ds-button/ds-button.tsx
@@ -1,8 +1,7 @@
 import { define } from '../preactement/define';
 import Button, { ButtonProps } from '../../Button/Button';
 import { analyticsAttrs } from '../shared-attributes/analytics';
-import { AnalyticsEvent } from '../../analytics';
-import { config } from '../../config';
+import { onAnalyticsEvent } from '../analytics';
 
 const attributes = [
   'class-name',
@@ -35,35 +34,4 @@ const Wrapper = ({ isAlternate, isOnDark, analytics, ...otherProps }: WrapperPro
   />
 );
 
-// What if we called config().defaultAnalyticsEvent within our own listener in here if
-// the user does not preventDefault()? Or maybe if some boolean is provided?
-
-define('ds-button', () => Wrapper, {
-  attributes,
-  events: [
-    'onClick',
-    [
-      'onAnalyticsEvent',
-      (event: AnalyticsEvent) => {
-        let callDefaultAnalyticsFunction = true;
-
-        function preventDefault() {
-          callDefaultAnalyticsFunction = false;
-        }
-
-        // Give event listeners a chance to prevent default before we execute the default
-        // analytics function from the config.
-        setTimeout(() => {
-          if (callDefaultAnalyticsFunction) {
-            config().defaultAnalyticsFunction(event);
-          }
-        }, 10);
-
-        return {
-          detail: { event },
-          preventDefault,
-        };
-      },
-    ],
-  ],
-});
+define('ds-button', () => Wrapper, { attributes, events: ['onClick', onAnalyticsEvent] });
